Fixed -Wconversion warnings

+++ src/champsim.cc
@@ -74,7 +74,7 @@ phase_stats do_phase(phase_info phase, environment& env, std::vector<tracereader
 
     // Read from trace
     for (O3_CPU& cpu : env.cpu_view())
-      std::generate_n(std::back_inserter(cpu.input_queue), cpu.IN_QUEUE_SIZE - std::size(cpu.input_queue), std::ref(traces.at(trace_index.at(cpu.cpu))));
+      std::generate_n(std::back_inserter(cpu.input_queue), cpu.IN_QUEUE_SIZE - static_cast<long>(std::size(cpu.input_queue)), std::ref(traces.at(trace_index.at(cpu.cpu))));
 
     // Check for phase finish
     auto [elapsed_hour, elapsed_minute, elapsed_second] = elapsed_time();
